Modules

+++ src/main.rs
@@ -1,6 +1,7 @@
 mod autoscaling;
 mod ecs;
 mod elasticache;
+mod rds;
 
 use clap::Parser;
 use log::{debug, info};
@@ -9,9 +10,9 @@ use tokio;
 #[derive(Parser)]
 #[command(author, version, about, long_about = None)]
 #[clap(
-    version = "v0.0.1",
+    version = "v0.1.0",
     author = "Anton Sidorov tonysidrock@gmail.com",
-    about = "Counts wwords frequency in a text file"
+    about = "Scale down ECS cluster"
 )]
 struct Args {
     #[clap(short, long, default_value = "direc-prod-lb")]
@@ -19,6 +20,12 @@ struct Args {
 
     #[clap(short, long, default_value = "eu-central-1")]
     region: String,
+
+    #[clap(short, long, default_value = "false")]
+    delete: bool,
+
+    #[clap(short, long, default_value = "false")]
+    scaledown: bool,
 }
 
 #[tokio::main]
@@ -31,25 +38,42 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let asgs = autoscaling::list_asgs(&as_client, &args.cluster, 0).await?;
     info!("ASGs: {:?}", asgs);
 
-    for asg in asgs {
-        autoscaling::scale_down_asg(&as_client, &asg, 0).await?;
-    }
-
     let elc_client = elasticache::initialize_client(&args.region).await;
     let replication_groups =
         elasticache::list_replication_groups(&elc_client, &args.cluster).await?;
     info!("Replication Groups: {:?}", replication_groups);
 
-    for replication_group in replication_groups {
-        elasticache::delete_replication_group(&elc_client, &replication_group).await?;
-    }
-
     let ecs_client = ecs::initialize_client(&args.region).await;
     let services = ecs::get_service_arns(&ecs_client, &args.cluster, 0).await?;
     info!("Services: {:?}", services);
 
-    for service in services {
-        ecs::scale_down_service(&ecs_client, &args.cluster, &service, 0).await?;
+    let rds_client = rds::initialize_client(&args.region).await;
+    let db_instances = rds::list_db_instances(&rds_client, &args.cluster).await?;
+    info!("DB Instances: {:?}", db_instances);
+
+    if args.scaledown || args.delete {
+        for asg in &asgs {
+            autoscaling::scale_down_asg(&as_client, &asg, 0).await?;
+        }
+        for service in &services {
+            ecs::scale_down_service(&ecs_client, &args.cluster, &service, 0).await?;
+        }
+        for db_instance in &db_instances {
+            rds::stop_db_instance(&rds_client, &db_instance).await?;
+        }
+    }
+
+    if args.delete {
+        for replication_group in replication_groups {
+            elasticache::delete_replication_group(&elc_client, &replication_group).await?;
+        }
+        for service in &services {
+            ecs::delete_service(&ecs_client, &args.cluster, &service).await?;
+        }
+        for db_instance in &db_instances {
+            rds::disable_deletion_protection(&rds_client, &db_instance).await?;
+            rds::delete_db_instance(&rds_client, &db_instance).await?;
+        }
     }
 
     debug!("Cluster: {} Region: {}.", &args.cluster, &args.region);
